moved Conquest map files duo in MapLoader duo as per requirements
